--- components/about-panel/index.ts
@@ -1,1 +1,0 @@
-export * from "./about-panel";
