+++ drizzle/seed-parts/string2double-struck.ts
